Initial Commit - Added Nunit test project -Added Interfaces

+++ SynetecAssessmentApi/Dtos/CalculateBonusDto.cs
@@ -2,7 +2,7 @@
 {
     public class CalculateBonusDto
     {
-        public int TotalBonusPoolAmount { get; set; }
+        public decimal TotalBonusPoolAmount { get; set; }
         public int SelectedEmployeeId { get; set; }
     }
 }
